+++ src/components/Modal.js
@@ -1,83 +1,19 @@
-import '../../styles/modal.css'
-import { closeModal } from '../../actions/modal'
-import { deleteRepl } from '../../actions/editor'
-import { connect } from 'react-redux'
-import { map } from 'lodash';
-import { RENDR_MODAL } from '../../utils/localStorageKeys'
-import { replace } from 'lodash'
-import Fade from './Fader'
-import PropTypes from 'prop-types'
+import '../styles/modal.css'
+
 import React, { Component } from 'react'
+
+import AnnouncementModal from './modals/AnnouncementModal';
+import BindingsModal from './modals/BindingsModal';
+import ConfirmModal from './modals/ConfirmModal';
+import Fade from './utils/Fader'
+import PropTypes from 'prop-types'
 import ReactDOM from 'react-dom'
+import ResourcesModal from './modals/ResourcesModal';
+import ThemeModal from './modals/ThemeModal';
+import { closeModal } from '../actions/modal'
+import { connect } from 'react-redux'
+import { deleteRepl } from '../actions/editor'
 import shortid from 'shortid'
-
-// modal types:
-const ConfirmModal = ({ closeModal, deleteRepl, id, theme }) => (
-  <div className="modal">
-    <h2 className="modal--header">
-      { `Are you sure you want to delete the "${replace(id, /_/g, ' ')}" repl?` }
-    </h2>
-    <div className="modal--confirm-buttons">
-      <div
-        className={`modal--confirm--yes ${theme}`}
-        id={'DELETE__' + replace(id, /\s/g, '_')}
-        onClick={deleteRepl}>
-        Yes
-      </div>
-      <div
-        className={`modal--confirm--no ${theme}`}
-        onClick={closeModal}>
-        No
-      </div>
-    </div>
-  </div>
-)
-
-const ResourcesModal = (props) => (
-  <div className="modal">
-    <h2 className="modal--header">
-      { props.header }
-    </h2>
-    <ul>
-      { map(props.messages, props.renderListItem) }
-    </ul>
-  </div>
-)
-
-const AnnouncementModal = (props) => (
-  <div className="modal">
-    <h2 className="modal--header">
-      { props.header }
-    </h2>
-    <p>
-      <strong dangerouslySetInnerHTML={{ __html: props.subHeader }} />
-    </p>
-    <ul>
-      { map(props.messages, props.renderListItem) }
-    </ul>
-    <p>
-      <strong dangerouslySetInnerHTML={{ __html:
-          props.messages[props.messages.length-1]}} />
-    </p>
-    {(() => {
-        let num = localStorage.getItem(RENDR_MODAL)
-        num = num === '1' ? 2 : num === '2' ? 1 : 0
-        return (
-          <span>
-            {`You will see this notification ${num} more time${num === 1 ? '' : 's'}`}
-          </span>
-        )
-    })()}
-  </div>
-)
-
-const ThemeModal = ({ header }) => (
-  <h2
-    className="modal"
-    style={{ padding: 20, boxShadow: 'none', opacity: 0.8 }}>
-    { header }
-  </h2>
-)
 
 class Modal extends Component {
   componentDidMount() {
@@ -144,6 +80,8 @@ class Modal extends Component {
               { ...this.props }
               renderListItem={this.renderListItem}
               renderNumAnnounced={this.renderNumAnnounced} />
+          : modalType === 'bindings'
+          ? <BindingsModal { ...this.props } />
           : modalType === 'resources'
           ? <ResourcesModal
               { ...this.props }