+++ src/components/utils/Fader.js
@@ -1,6 +1,7 @@
 // @flow
-import { Transition } from 'react-transition-group'
 import * as React from 'react'
+
+import { Transition } from 'react-transition-group'
 
 const transitionStyles: { [string]: Object } = {
   entering: { opacity: 0 },
@@ -12,18 +13,19 @@ type Props = {
   in: boolean,
   children: React.Node,
   attachRef: Function,
-  duration: number | { enter: number, exit: number }
+  duration: string | { enter: string, exit: string }
 }
 
 const Fade = ({ in: inProp, children, attachRef, duration }: Props) => {
   const defaultStyle = {
-    transition: `opacity ${duration.exit ? duration.exit : duration}ms ease-out`,
+    // $FlowFixMe = stupid coercion problem
+    transition: `opacity ${duration.exit ? duration.exit : duration} ms ease-out`,
     opacity: 0
   }
   return (
     <Transition in={inProp} timeout={duration} unmountOnExit={true}>
       {(state) => (
-        <div ref={attachRef} style={{
+        <div ref={attachRef && attachRef} style={{
           ...defaultStyle,
           ...transitionStyles[state]
         }}>